requirements.txt
# requirements.txt
faker==19.6.2
kafka-python==2.0.2
redis==4.6.0
apache-airflow==2.7.1
pyspark==3.4.1
tensorflow==2.13.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0